Initial code to manage context mangling.
Chris Johns mentioned that Stuart Swales had spotted bad use of
context handles in Unixlib in the filesystem enumeration. There's
usually problems with filesystem enumeration when people fail to
handle the 'no values returned but we're still enumerating' or the
'some values returned and we're no longer enumerating' cases, but
manipulating the opaque context value is a definite no-no.
Filer_Action is also known to be evil and do this.

These classes here are intended to make the opaque handles returned
use a different pattern to that which you might expect - there are
a few different ways in which the enumeration can be modified such
as making it go down, or even semi-randomly by changing the bit
order.

There's only changes here which handle the mangling, not any
integration with RISC OS itself.